refactor: Attachment, Assignment Upload facade -> service, SubmitApiControllerTest 모킹 오류 수정
