+++ src/main/java/com/kgu/studywithme/studyweekly/application/service/AttachmentUploader.java
@@ -1,4 +1,4 @@
-package com.kgu.studywithme.studyweekly.application.facade;
+package com.kgu.studywithme.studyweekly.application.service;
 
 import com.kgu.studywithme.file.application.adapter.FileUploader;
 import com.kgu.studywithme.studyweekly.domain.attachment.UploadAttachment;
